src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/SelectorAPI/only-shadow-host-in-shadow-tree-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
crbug.com/359854: test for traversing elements in shadow tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
